requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
astropy==6.1.0
astropy-iers-data==0.2024.5.27.0.30.8
bcrypt==4.1.3
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
docstring_parser==0.16
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
orjson==3.10.3
packaging==24.0
passlib==1.7.4
pluggy==1.5.0
psycopg2==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
pyerfa==2.0.1.4
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.2.1
python-dotenv==1.0.1
python-multipart==0.0.9
pyuvdata==2.4.3
PyYAML==6.0.1
rich==13.7.1
scipy==1.13.1
setuptools==70.0.0
setuptools-scm==8.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
